AK-94 [Feat] 사용자 권한 검증 로직 추가 및 DB 수정, 삭제기능 수정
