+++ fastapi/requirements.txt
@@ -1,0 +1,26 @@
+beautifulsoup4==4.12.3
+boto3==1.35.63
+easyocr==1.7.2
+elasticsearch==8.16.0
+fastapi==0.115.5
+nest_asyncio==1.6.0
+numpy==1.24.3
+openai==1.55.0
+opencv_contrib_python==4.10.0.84
+opencv_python==4.10.0.84
+opencv_python_headless==4.10.0.84
+pandas==2.0.3
+Pillow==11.0.0
+pydantic==2.10.0
+pymysql==1.1.1
+python_Levenshtein==0.25.1
+python_Levenshtein==0.26.1
+pytube==15.0.0
+Requests==2.32.3
+scikit_learn==1.3.0
+SQLAlchemy==2.0.34
+starlette==0.41.3
+torch==2.2.2
+uvicorn==0.32.1
+youtube_transcript_api==0.6.3
+yt_dlp==2024.11.18
